+++ packages/wallet/wallet-toolbox/src/storage/remoting/__test/StorageClient.test.ts
@@ -34,7 +34,8 @@ describe('StorageClient tests', () => {
     client = await _tu.createTestWalletWithStorageClient({
       rootKeyHex: server.setup.rootKey.toHex(),
       endpointUrl: 'http://localhost:8042',
-      chain: server.setup.chain
+      chain: server.setup.chain,
+      actionBatchMode: 'auto'
     })
     attacker = await _tu.createTestWalletWithStorageClient({
       rootKeyHex: '2'.repeat(64),